test(libdoc): split libdoc-builder by behaviour family
Split libdoc-builder.test.js (477 LOC) into libdoc-builder-pages.test.js
(PagesBuilder construction + rendering) and libdoc-builder-server.test.js
(PagesServer construction + serving) along the class boundary. Assertions
unchanged; 7 tests preserved.

libdoc-llms.test.js is retained over the ceiling on the allow-list — its
six tests all exercise the single llms.txt augmentation behaviour with no
clean sub-family seam.

Spec 0640 § C (large files), design Decision 5.
